requirements.txt
astroid==2.4.2
backoff==1.10.0
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
idna==2.10
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
ratelimit==2.2.1
requests==2.25.0
six==1.15.0
tinydb==4.3.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
